Implement Smalltalk's `do:` method
I had refrained from implementing this at first, because it's basically
the same thing as `enumerateObjectsUsingBlock:` (available starting with
10.6). But I decided to implement it because `do:` takes a simpler block
than `enumerateObjectsUsingBlock:`, I can return `self` in `do:` so
several calls can be chained together, and, well, libCollections wouldn't
be emulating the Smalltalk API without `do:`, now would it?

Plus I'd like to eventually compile libCollections on 10.5, so this
would allow some 10.6 functionality to be pseudo-backported to 10.5.