Use build number for compatibility information
According to the Apple docs, "the current version number tracks
individual builds of your framework and is mostly for internal use
by your team. [...] When you make certain kinds of changes to public
interfaces, you should set the compatibility version to match the
current version of your framework." So I'm going to use the
$CURRENT_PROJECT_VERSION to record compatibility information for the
library; I'll increment this number whenever I create a build for
internal or external testing or external consumption. This will be
maintained separately from the "marketing" version of the library,
which will be the more familiar "1.0", "1.0.1", etc.