Build static library for iOS 4.2
The library is built for both iPhone and iPad.

As expected, no changes were required to compile for iOS. However, I had
to make a separate project file; it is my understanding that the same
project file cannot contain targets for both Mac OS X and iOS.
Disappointing, because each project file will now have to be kept in sync
as files are added and deleted, but that's how it is.

I haven't actually tested this in any iOS apps, but again, the library
is pretty simple so it should work fine.