Replace common macros with methods
This might be a bit "uglier" than before in some ways, but most of the
ugliness won't be exposed to the user. The only problem I have is with
the method `setObject:forValue`, which seems likely to run into
conflicts (and sounds a lot like an existing method in the Foundation
framework). I may change the name of this method eventually.